Allow more decimal places when adding products
The product form's number inputs used step="0.01" (prices) and step="1"
(opening stock, reorder level), so the browser rejected values like 1.235
or 2.5 before submitting. They now accept any decimals.

Prices/costs (products and variants) are stored to 6 decimal places (was 4),
and reorder level / reorder quantity / alert quantity become 3-decimal
numbers (were whole numbers) so kg-sold products can use e.g. 2.5; the
validation and Excel import no longer truncate them.

+++ backend/app/Http/Controllers/Api/InventoryController.php
@@ -374,7 +374,7 @@ class InventoryController extends BaseApiController
                         'category_id'   => $categoryId,
                         'cost_price'    => $costPrice ?? 0,
                         'selling_price' => $sellingPrice ?? 0,
-                        'reorder_level' => !empty($row['reorder_level']) ? intval($row['reorder_level'])   : 5,
+                        'reorder_level' => !empty($row['reorder_level']) ? floatval($row['reorder_level'])   : 5,
                         'track_stock'   => true,
                         'is_active'     => true,
                     ];
